Consolidate fetching livestream info
