+++ src/Livestream/LivestreamInfo.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mati\Livestream;
+
+final readonly class LivestreamInfo
+{
+  public function __construct(
+    public string $livestreamUrl,
+    public string $chatUrl,
+    public int $chatId,
+  ) {
+    // Do nothing.
+  }
+}
